Extracting `useLogin` hook.
